requirements.txt
autopep8==1.7.0
grpcio==1.50.0
grpcio-tools==1.50.0
protobuf==4.21.8
pycodestyle==2.9.1
six==1.16.0
toml==0.10.2
